# swaggerDocumentation.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Artworks"
  },
  "servers": [
    {
      "url": "https://blooming-cliffs-86757.herokuapp.com/"
    }
  ],
  "paths": {
    "/artworks": {
      "get": {
        "summary": "List all artworks or the artwork which created by the particular artist",
        "tags": [
          "artworks"
        ],
        "parameters": [
          {
            "name": "artist",
            "in": "query",
            "required": false,
            "description": "Filters artwork that is created by the supplied query artist name.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of the artist's artwork",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "example": {
                      "artwork": "orange container of nagpur",
                      "artist": "bijay biswaal",
                      "type": "watercolor",
                      "subject": "cityscape",
                      "surface": "paper",
                      "size": "14x22 inch",
                      "description": "painted by bijay biswaal",
                      "price": "583.33",
                      "image_url": "https://www.artzolo.com/sites/default/files/uploads/multi/101/largest/orange_container_of_nagpur.jpg"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new artwork",
        "tags": [
          "artworks"
        ],
        "requestBody": {
          "description": "the required fields for creating the artwork",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "artwork": "artwork name",
                  "artist": "artist name",
                  "size": "the dimension of the artwork in inch",
                  "price": "the price of the artwork"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expected the new artwork is contained the required field",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "example": {
                      "artwork": "orange container of nagpur",
                      "artist": "bijay biswaal",
                      "size": "14x22 inch",
                      "price": "583.33"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/artworks/id": {
      "put": {
        "summary": "Updated fields for specific artwork",
        "tags": [
          "artworks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The artwork id which to be updated",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "the fields for artwork which to be updated",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "price": 2000
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Expected the artwork id valid",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "5b4ff61a0550226fb25a8653",
                    "artwork": "banaras ganga ghat 15",
                    "artist": "ganesh hire",
                    "type": "watercolor",
                    "subject": "nature",
                    "surface": "paper",
                    "size": "22x45 inch",
                    "description": "",
                    "amount": 2000,
                    "image_url": "https://www.artzolo.com/sites/default/files/uploads/multi/4964/largest/22x45_d.jpg"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Update a specific artwork detail",
        "tags": [
          "artworks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The artwork of the id to be removed",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Expected the artwork id is valid",
            "content": {
              "application/json": {
                "schema": {
                  "example": {}
                }
              }
            }
          }
        }
      }
    }
  }
}